requirements.txt
connected_components_3d==3.20.0
matplotlib==3.8.3
numpy==1.26.3
opencv_python==4.11.0.86
segmentation_models_pytorch==0.4.0
# Install torch manually
# torch==2.3.0+cu121
